Split up index.php
