+++ backend/lib/results/IResult.php
@@ -1,0 +1,6 @@
+<?php
+namespace Lib\Results;
+
+interface IResult {
+    public function execute();
+}
