+++ backend/lib/results/JSON.php
@@ -1,0 +1,15 @@
+<?php
+namespace Lib\Results;
+
+class JSON implements IResult {
+    private $obj;
+
+    function __construct($obj) {
+        $this->obj = $obj;
+    }
+
+    public function execute() {
+        header('Content-type: application/json');
+        echo json_encode($this->obj);
+    }
+}
